Reflexion Agent/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
grandalf==0.8
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
isort==6.0.1
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.55
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langchain-tavily>=0.1.0
langgraph==0.3.33
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.63
langsmith==0.3.33
mypy-extensions==1.1.0
openai==1.76.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pyparsing==3.2.3
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlalchemy==2.0.40
tenacity==9.1.2
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
xxhash==3.5.0
zstandard==0.23.0 